+++ crates/gdtk-gdscript-parser/src/expressions.rs
@@ -1,7 +1,7 @@
 use gdtk_ast::{
     ASTBinaryOp, ASTExpr, ASTExprKind, ASTPostfixOp, ASTPostfixOpKind, ASTPrefixOp, ASTPrefixOpKind,
 };
-use gdtk_lexer::{Token, TokenKind};
+use crate::lexer::{Token, TokenKind};
 use gdtk_span::Span;
 use pratt::{Affix, Associativity, PrattParser, Precedence};
