+++ crates/gdtk-gdscript-parser/src/values.rs
@@ -1,5 +1,5 @@
 use gdtk_ast::{ASTExpr, ASTFunction, ASTFunctionKind, DictValue};
-use gdtk_lexer::{Token, TokenKind};
+use crate::lexer::{Token, TokenKind};
 
 use crate::{
     expressions::parse_expr,
