--- crates/gdtk-parser/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "gdtk-parser"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-gdtk-span = { path = "../gdtk-span", version = "0.1.0" }
-gdtk-lexer = { version = "0.1.0", path = "../gdtk-lexer" }
-gdtk-ast = { version = "0.1.0", path = "../gdtk-ast" }
-pratt = "0.4.0"
-enum-as-inner = "0.6.0"
