+++ crates/gdtk/src/commands/dev/lex.rs
@@ -6,7 +6,7 @@ use crate::utils::get_content;
 
 pub fn run(file: PathBuf) -> anyhow::Result<()> {
     let content = get_content(file.as_path())?;
-    let lexed = gdtk_lexer::lex(&content);
+    let lexed = gdtk_gdscript_parser::lexer::lex(&content);
 
     eprintln!("Lexer output:\n```ron\n{:#?}\n```", &lexed.collect_vec());
