+++ crates/gdtk/src/commands/dev/parse.rs
@@ -4,8 +4,8 @@ use crate::utils::get_content;
 
 pub fn run(file: PathBuf) -> anyhow::Result<()> {
     let content = get_content(file.as_path())?;
-    let lexed = gdtk_lexer::lex(&content);
-    let parsed = gdtk_parser::parse_file(lexed);
+    let lexed = gdtk_gdscript_parser::lexer::lex(&content);
+    let parsed = gdtk_gdscript_parser::parse_file(lexed);
 
     eprintln!("Parser output:\n```ron\n{:#?}\n```", &parsed);
